- renames the blank fragment to root fragment - changes splash logo
