+++ app/src/main/res/layout/fragment_root.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".BlankFragment">
+    tools:context=".root.presentation.RootFragment">
 
     <androidx.viewpager2.widget.ViewPager2
         android:id="@+id/vp2"
