+++ app/src/main/res/navigation/job_pulse_nav.xml
@@ -3,7 +3,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/job_pulse_nav"
-    app:startDestination="@id/blankFragment">
+    app:startDestination="@id/rootFragment">
 
     <fragment
         android:id="@+id/filters"
@@ -94,19 +94,19 @@
         tools:layout="@layout/fragment_district" />
 
     <fragment
-        android:id="@+id/blankFragment"
-        android:name="ru.practicum.android.diploma.BlankFragment"
-        android:label="fragment_blank"
-        tools:layout="@layout/fragment_blank">
+        android:id="@+id/rootFragment"
+        android:name="ru.practicum.android.diploma.root.presentation.RootFragment"
+        android:label="fragment_root"
+        tools:layout="@layout/fragment_root">
         <action
-            android:id="@+id/action_blankFragment_to_filters"
+            android:id="@+id/action_rootFragment_to_filters"
             app:destination="@id/filters"
             app:enterAnim="@anim/slide_in_from_right"
             app:exitAnim="@anim/slide_out_left"
             app:popEnterAnim="@anim/slide_in_left"
             app:popExitAnim="@anim/slide_out_right"/>
         <action
-            android:id="@+id/action_blankFragment_to_vacancy"
+            android:id="@+id/action_rootFragment_to_vacancy"
             app:destination="@id/vacancy"
             app:enterAnim="@anim/slide_in_from_right"
             app:exitAnim="@anim/slide_out_left"
